- first sketch of graph capabilities. graph will allow arbitrary vertex types with a VertexLike mixin and a single defined Edge type. edges are not weightened or anything but store flow information (see EdgeKind)
the backend of a graph is modeled by a trait -- so we might be able to switch the data structure for sparse/dense/tree graphs etc. the inheritance graph of a swf should be a tree for instance so a different structure might be more appropriate

dsl syntax is provided as well. G += a -> b works, however in order to perform pattern matching on might do G.incommingOf(a)(0) match { case c ~> d => ... }

we will see if the pattern matching stays like that.